requirements.txt
blinker==1.9.0
click==8.1.8
exceptiongroup==1.3.0
Flask==3.1.1
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-flask==1.3.0
tomli==2.2.1
typing_extensions==4.14.0
Werkzeug==3.1.3
zipp==3.23.0
